+++ src/decode.ts
@@ -1,7 +1,9 @@
 import { MessageTypeNames, MessageTypes } from './message-types.js'
 import { Uint8ArrayList } from 'uint8arraylist'
-import type { Source } from 'it-stream-types'
 import type { Message } from './message-types.js'
+
+export const MAX_MSG_SIZE = 1 << 20 // 1MB
+export const MAX_MSG_QUEUE_SIZE = 4 << 20 // 4MB
 
 interface MessageHeader {
   id: number
@@ -13,10 +15,14 @@ interface MessageHeader {
 export class Decoder {
   private readonly _buffer: Uint8ArrayList
   private _headerInfo: MessageHeader | null
+  private readonly _maxMessageSize: number
+  private readonly _maxUnprocessedMessageQueueSize: number
 
-  constructor () {
+  constructor (maxMessageSize: number = MAX_MSG_SIZE, maxUnprocessedMessageQueueSize: number = MAX_MSG_QUEUE_SIZE) {
     this._buffer = new Uint8ArrayList()
     this._headerInfo = null
+    this._maxMessageSize = maxMessageSize
+    this._maxUnprocessedMessageQueueSize = maxUnprocessedMessageQueueSize
   }
 
   write (chunk: Uint8Array) {
@@ -25,13 +31,22 @@ export class Decoder {
     }
 
     this._buffer.append(chunk)
+
+    if (this._buffer.byteLength > this._maxUnprocessedMessageQueueSize) {
+      throw Object.assign(new Error('unprocessed message queue size too large!'), { code: 'ERR_MSG_QUEUE_TOO_BIG' })
+    }
+
     const msgs: Message[] = []
 
     while (this._buffer.length !== 0) {
       if (this._headerInfo == null) {
         try {
           this._headerInfo = this._decodeHeader(this._buffer)
-        } catch (_) {
+        } catch (err: any) {
+          if (err.code === 'ERR_MSG_TOO_BIG') {
+            throw err
+          }
+
           break // We haven't received enough data yet
         }
       }
@@ -81,6 +96,11 @@ export class Decoder {
       throw new Error(`Invalid type received: ${type}`)
     }
 
+    // test message type varint + data length
+    if (length > this._maxMessageSize) {
+      throw Object.assign(new Error('message size too large!'), { code: 'ERR_MSG_TOO_BIG' })
+    }
+
     // @ts-expect-error h is a number not a CODE
     return { id: h >> 3, type, offset: offset + end, length }
   }
@@ -113,20 +133,5 @@ function readVarInt (buf: Uint8ArrayList, offset: number = 0) {
   return {
     value: res,
     offset
-  }
-}
-
-/**
- * Decode a chunk and yield an _array_ of decoded messages
- */
-export async function * decode (source: Source<Uint8Array>) {
-  const decoder = new Decoder()
-
-  for await (const chunk of source) {
-    const msgs = decoder.write(chunk)
-
-    if (msgs.length > 0) {
-      yield msgs
-    }
   }
 }